+++ examples/post-analysis-tree/content/chapter/topic/extension/mark.lua
@@ -1,0 +1,6 @@
+function Header(header)
+  header.content:insert(pandoc.Space())
+  header.content:insert(pandoc.Str("filtered"))
+  return header
+end
+